requirements.txt
# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# UI
customtkinter==5.2.1
ttkthemes==3.2.2
pillow==10.0.1

# LLM Integration
anthropic==0.5.0       # For Claude models
transformers==4.35.2   # For Hugging Face models
torch==2.1.1
sentencepiece==0.1.99
huggingface_hub==0.19.4
langchain==0.0.335     # Framework for working with LLMs

# PDF Generation
reportlab==4.0.4
matplotlib==3.7.3
seaborn==0.13.0
pandas==2.1.1

# Utilities
python-dotenv==1.0.0
nltk==3.8.1
textblob==0.17.1       # Simple sentiment analysis